moved *SubSwitch classes to package subswitches
