+++ org.opentosca.yamlconverter.main/src/main/java/org/opentosca/yamlconverter/switchmapper/subswitches/CapabilityTypesSubSwitch.java
@@ -1,6 +1,7 @@
-package org.opentosca.yamlconverter.switchmapper;
+package org.opentosca.yamlconverter.switchmapper.subswitches;
 
 import org.opentosca.model.tosca.TCapabilityType;
+import org.opentosca.yamlconverter.switchmapper.Yaml2XmlSwitch;
 import org.opentosca.yamlconverter.switchmapper.typemapper.ElementType;
 import org.opentosca.yamlconverter.yamlmodel.yaml.element.CapabilityType;
