+++ org.opentosca.yamlconverter.main/src/main/java/org/opentosca/yamlconverter/switchmapper/subswitches/RelationshipTypesSubSwitch.java
@@ -1,7 +1,8 @@
-package org.opentosca.yamlconverter.switchmapper;
+package org.opentosca.yamlconverter.switchmapper.subswitches;
 
 import org.opentosca.model.tosca.TInterface;
 import org.opentosca.model.tosca.TRelationshipType;
+import org.opentosca.yamlconverter.switchmapper.Yaml2XmlSwitch;
 import org.opentosca.yamlconverter.switchmapper.typemapper.ElementType;
 import org.opentosca.yamlconverter.yamlmodel.yaml.element.RelationshipType;
